+++ src/extensions/row_extension.rs
@@ -1,20 +1,20 @@
 use crate::models::csv_config::CsvConfig;
 use crate::models::in_row_iter::InRowIter;
 
-pub trait IterableRow<'a> {
+pub(crate) trait IterableRow<'mmap> {
     /// ## Get Iterator
     ///
     /// Extracts the `InRowIter` struct.
     /// This struct, can move inside the raw row, without copy or allocation.
     /// - `returns`: `InRowIter<'a>`
-    fn get_iterator(self, cfg: &CsvConfig) -> InRowIter<'a>;
+    fn get_iterator(self, cfg: &CsvConfig) -> InRowIter<'mmap>;
 }
 
 
 //Implementation of the trait
 
-impl<'a> IterableRow<'a> for &'a [u8] {
-    fn get_iterator(self, cfg: &CsvConfig) -> InRowIter<'a> {
+impl<'mmap> IterableRow<'mmap> for &'mmap [u8] {
+    fn get_iterator(self, cfg: &CsvConfig) -> InRowIter<'mmap> {
         InRowIter::new(
             self,
             cfg.delimiter,
@@ -22,3 +22,5 @@ impl<'a> IterableRow<'a> for &'a [u8] {
         )
     }
 }
+
+