+++ src/models/mod.rs
@@ -4,3 +4,6 @@ pub mod csv_error;
 pub mod datatype;
 pub mod platform_info;
 pub mod in_row_iter;
+pub mod row;
+pub mod workers;
+mod field;